Move the sources package back into the resolver package
