+++ packages/resolver/sources/globalnpm.js
