+++ packages/resolver/sources/npm.js
