--- packages/sources/package.json
@@ -1,22 +1,0 @@
-{
-  "name": "sources",
-  "version": "1.0.0",
-  "description": "",
-  "main": "index.js",
-  "dependencies": {
-    "debug": "^3.2.6",
-    "detect-installed": "^2.0.4",
-    "get-installed-path": "^4.0.8",
-    "glob": "^7.1.3",
-    "supports-color": "^6.0.0"
-  },
-  "devDependencies": {
-    "mocha": "^6.2.2"
-  },
-  "scripts": {
-    "test": "mocha"
-  },
-  "keywords": [],
-  "author": "",
-  "license": "ISC"
-}
